+++ src/libs/pr.ts
@@ -1,8 +1,48 @@
 import * as core from '@actions/core';
 import { context, getOctokit } from '@actions/github';
-import * as dedent from 'dedent';
+import AnsiToHtml from 'ansi-to-html';
+import dedent from 'dedent';
 import invariant from 'ts-invariant';
 import { Config } from '../config';
+
+function ansiToHtml(
+  message: string,
+  maxLength: number,
+  alwaysIncludeSummary: boolean,
+): [string, boolean] {
+  /**
+   *  Converts an ansi string to html by for example removing color escape characters.
+   *  message: ansi string to convert
+   *  maxLength: Maximum number of characters of final message incl. HTML tags
+   *  alwaysIncludeSummary: if true, trim message from front (if trimming is needed), otherwise from end
+   *
+   *  return message as html and information if message was trimmed because of length
+   */
+  const convert = new AnsiToHtml();
+  let trimmed = false;
+
+  let htmlBody: string = convert.toHtml(message);
+
+  // Check if htmlBody exceeds max characters
+  if (htmlBody.length > maxLength) {
+
+    // trim input message by number of exceeded characters from front or back as configured
+    const dif: number = htmlBody.length - maxLength;
+
+    if (alwaysIncludeSummary) {
+      message = message.substring(dif, htmlBody.length);
+    } else {
+      message = message.substring(0, message.length - dif);
+    }
+
+    trimmed = true;
+
+    // convert trimmed message to html
+    htmlBody = convert.toHtml(message);
+  }
+
+  return [htmlBody, trimmed];
+}
 
 export async function handlePullRequestMessage(
   config: Config,
@@ -14,28 +54,34 @@ export async function handlePullRequestMessage(
     command,
     stackName,
     editCommentOnPr,
+    alwaysIncludeSummary,
   } = config;
+
+  // GitHub limits comment characters to 65535, use lower max to keep buffer for variable values
+  const MAX_CHARACTER_COMMENT = 64_000;
 
   const heading = `#### :tropical_drink: \`${command}\` on ${projectName}/${stackName}`;
 
   const summary = '<summary>Pulumi report</summary>';
 
-  const rawBody = output.substring(0, 64_000);
-  // a line break between heading and rawBody is needed
-  // otherwise the backticks won't work as intended
+  const [htmlBody, trimmed]: [string, boolean] = ansiToHtml(output, MAX_CHARACTER_COMMENT, alwaysIncludeSummary);
+
   const body = dedent`
     ${heading}
 
     <details>
     ${summary}
 
-    \`\`\`
-    ${rawBody}
-    \`\`\`
-    ${
-      rawBody.length === 64_000
-        ? '**Warn**: The output was too long and trimmed.'
-        : ''
+    ${trimmed && alwaysIncludeSummary
+      ? ':warning: **Warn**: The output was too long and trimmed from the front.'
+      : ''
+    }
+    <pre>
+    ${htmlBody}
+    </pre>
+    ${trimmed && !alwaysIncludeSummary
+      ? ':warning: **Warn**: The output was too long and trimmed.'
+      : ''
     }
     </details>
   `;